Updates a directory to be included in project-config
When this spec was written nodepool had one directory beneath,
the scripts directory. There is a patch up which will create
a nodepool/elements directory. If that patch lands before
the config repo is split, the elements directory should travel
with the nodepool directory and live in project-config.

Patch that creates the elements directory:
https://review.openstack.org/#/c/88479/

Change-Id: I9119c1ca81d1f4a48d99f17c17ffdc25d834670b