Adds an implemented directory
Right now infra-specs doesn't have a way to indicate which specs
have been implemented. This patch adds a directory entitled
implemented to gather completed specs. One spec which is known
to be implemented is moved into this directory.

The documentation index file has been edited so that the
documentation reflects this change.

Change-Id: I69be74b7469030abcbaf1c0e341b38e5053500c5