Host a code search service
It is often convenient, for instance when researching certain
classes of widespread bugs or looking for implementation examples,
to be able to search through all of the OpenStack projects' source
code. As a service to our contributors and downstream communities,
we should provide a convenient mechanism to be able to do this.

Change-Id: I2ee342b86dac092cc3024ae1b5a211a7335c7d3b
Story: #2000280