Zuul v3: use Zookeeper for Nodepool-Zuul protocol
This updates the Nodepool-Zuul protocol to use Zookeeper after the
fashion of the spec to use Zookeeper for Nodepool image builds.

Despite this having more words than the section it replaces, this
is ultimately a much more robust protocol (particularly around
race conditions and edge cases related to Gearman disconnections)
and should be simpler to implement in Nodepool.  Notably, this
removes the last central component of Nodepool allowing it to be
a truly decentralised (and much more fault-tolerant) app.

Change-Id: I354f3eb8de3a8218cea04f5b19bb4b05f4e0cb60