Overhaul instructions in README.rst for clarity
The README.rst file included some rough instructions which may not
be entirely clear to newcomers to the community. Add some details so
that they don't need to guess where some things are.

Switch the recommendation for Story commit footers to Task so that
the corresponding story task will get its status updated by our
automation accordingly (we hyperlink these since the SB webclient
has grown support for routing them to the correct story, so
including a Story footer as well is now unnecessary).

Drop the step of providing the review link in a story comment since
our automation will do this if a Task footer is included in the
commit message.

Change-Id: I1dcba7c88efa20b542f30f3f34a043caba7a4c3f